course-project/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pymongo
